+++ src/UserSearchModal.js
@@ -1,6 +1,6 @@
-import React, { Component } from 'react';
-import { FormattedMessage } from 'react-intl';
+import { useState, useRef, forwardRef, useEffect } from 'react';
 import PropTypes from 'prop-types';
+import { FormattedMessage } from 'react-intl';
 
 import { Modal } from '@folio/stripes/components';
 
@@ -9,122 +9,121 @@ import UserSearchView from './UserSearchView';
 
 import css from './UserSearch.css';
 
-class UserSearchModal extends Component {
-  static propTypes = {
-    stripes: PropTypes.shape({
-      connect: PropTypes.func.isRequired,
-      okapi: PropTypes.object.isRequired,
-    }).isRequired,
-    selectUsers: PropTypes.func,
-    selectUser: PropTypes.func,
-    closeCB: PropTypes.func.isRequired,
-    onCloseModal: PropTypes.func,
-    openWhen: PropTypes.bool,
-    dataKey: PropTypes.string,
-    contentRef: PropTypes.object,
-    modalRef: PropTypes.object,
-    modalTitle: PropTypes.node,
-    restoreFocus: PropTypes.bool,
-    initialSelectedUsers: PropTypes.shape({
-      [PropTypes.string]: PropTypes.shape({
-        username: PropTypes.string,
-        id: PropTypes.string,
-        active: PropTypes.bool,
-        barcode: PropTypes.string,
-        personal: PropTypes.shape({
-          lastName: PropTypes.string,
-          firstName: PropTypes.string,
-          email: PropTypes.string,
-        }),
-        patronGroup: PropTypes.string,
-      }),
-    }),
-    tenantId: PropTypes.string,
-  }
-
-  static defaultProps = {
-    restoreFocus: true,
-  };
-
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      error: null,
-    };
-
-    this.closeModal = this.closeModal.bind(this);
-    this.passUserOut = this.passUserOut.bind(this);
-    this.modalRef = props.modalRef || React.createRef();
-  }
-
-  closeModal() {
-    this.setState({
-      error: null,
-    }, () => {
-      this.props.closeCB();
-    });
-  }
-
-  passUserOut(e, user) {
-    this.props.selectUser(user);
-
-    if (!user.error) {
-      this.closeModal();
-    } else {
-      this.setState({
-        error: user.error,
-      });
-    }
-  }
-
-  passUsersOut = users => {
-    this.props.selectUsers(users);
-    this.closeModal();
-  }
-
-  render() {
-    const {
-      restoreFocus,
+const UserSearchModal = (
+  (
+    {
+      stripes,
+      selectUsers,
+      selectUser,
+      closeCB,
       onCloseModal,
       openWhen,
-      selectUsers,
+      restoreFocus = true,
       initialSelectedUsers,
       tenantId,
-      stripes,
-      modalTitle,
-    } = this.props;
+      ...rest
+    },
+    modalRef
+  ) => {
+    const [error, setError] = useState(null);
+    const [callbackTrigger, setCallbackTrigger] = useState(false);
+    const ref = useRef();
+    const mRef = modalRef || ref;
+
+    useEffect(() => {
+      if (callbackTrigger && error === null) {
+        closeCB();
+        setCallbackTrigger(false);
+      }
+    }, [callbackTrigger, error, closeCB]);
+
+    const closeModal = () => {
+      setError(null);
+      setCallbackTrigger(true); // Trigger the callback after state update
+    };
+
+    const passUserOut = (e, user) => {
+      selectUser(user);
+
+      if (!user.error) {
+        closeModal();
+      } else {
+        setError(user.error);
+      }
+    };
+
+    const passUsersOut = users => {
+      selectUsers(users);
+      closeModal();
+    };
 
     return (
       <Modal
         contentClass={css.modalContent}
         dismissible
         enforceFocus={false}
-        label={modalTitle || <FormattedMessage id="ui-plugin-find-user.modal.label" />}
+        label={<FormattedMessage id="ui-plugin-find-user.modal.label" />}
         open={openWhen}
-        ref={this.modalRef}
+        ref={mRef}
         size="large"
-        onClose={this.closeModal}
+        onClose={closeModal}
         restoreFocus={restoreFocus}
       >
-        {this.state.error ? <div className={css.userError}>{this.state.error}</div> : null}
+        {error ? <div className={css.userError}>{error}</div> : null}
         <UserSearchContainer
-          {...this.props}
+          stripes={stripes}
+          selectUsers={selectUsers}
+          selectUser={selectUser}
+          closeCB={closeCB}
+          onCloseModal={onCloseModal}
+          openWhen={openWhen}
+          restoreFocus={restoreFocus}
           onComponentWillUnmount={onCloseModal}
           tenantId={tenantId || stripes.okapi.tenant}
           initialSelectedUsers={initialSelectedUsers}
+          {...rest}
         >
           {(viewProps) => <UserSearchView
             {...viewProps}
-            onSaveMultiple={this.passUsersOut}
-            onSelectRow={this.passUserOut}
+            onSaveMultiple={passUsersOut}
+            onSelectRow={passUserOut}
             isMultiSelect={Boolean(selectUsers)}
             initialSelectedUsers={initialSelectedUsers}
           />}
         </UserSearchContainer>
       </Modal>
     );
-  }
-}
+  });
 
-export default UserSearchModal;
+UserSearchModal.propTypes = {
+  stripes: PropTypes.shape({
+    connect: PropTypes.func.isRequired,
+    okapi: PropTypes.object.isRequired,
+  }).isRequired,
+  selectUsers: PropTypes.func,
+  selectUser: PropTypes.func,
+  closeCB: PropTypes.func.isRequired,
+  onCloseModal: PropTypes.func,
+  openWhen: PropTypes.bool,
+  dataKey: PropTypes.string,
+  contentRef: PropTypes.object,
+  modalRef: PropTypes.object,
+  restoreFocus: PropTypes.bool,
+  initialSelectedUsers: PropTypes.shape({
+    [PropTypes.string]: PropTypes.shape({
+      username: PropTypes.string,
+      id: PropTypes.string,
+      active: PropTypes.bool,
+      barcode: PropTypes.string,
+      personal: PropTypes.shape({
+        lastName: PropTypes.string,
+        firstName: PropTypes.string,
+        email: PropTypes.string,
+      }),
+      patronGroup: PropTypes.string,
+    }),
+  }),
+  tenantId: PropTypes.string,
+};
+
+export default forwardRef(UserSearchModal);